Put scene modal in its own component
