+++ src/components/scene/sceneModal.css
@@ -1,0 +1,92 @@
+.melofi__scene-modal {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    top: 0;
+    right: 28px;
+    bottom: 0;
+    margin: auto;
+    background-color: var(--color-primary);
+    width: 300px;
+    height: 65%;
+    border-radius: 10px;
+    backdrop-filter: blur(10px);
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    padding: 15px 15px 0px 15px;
+    animation: scale-up 0.4s;
+}
+
+.melofi__scene-modal-title {
+    color: var(--color-secondary-white);
+    font-family: var(--font-poppins);
+    font-size: 21px;
+}
+
+.melofi__scene-modal-thumbnail-container{
+    display: flex;
+    flex-direction: column;
+    row-gap: 15px;
+    height: 90%;
+    margin-top: 10px;
+    overflow: scroll;
+}
+
+::-webkit-scrollbar {
+    width: 0px;
+    background: transparent; /* make scrollbar transparent */
+}
+
+.melofi__scene-modal-thumbnail {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    min-height: 250px;    
+    background-size:cover;
+    background-repeat: no-repeat;
+    background-position: center;
+    border-radius: 10px;
+    background-color: var(--color-primary);
+    opacity: 70%;
+    border: 1px solid var(--color-secondary-opacity);
+    align-items: center;
+    justify-content: center;
+}
+
+.melofi__scene-modal-thumbnail:hover {
+    opacity: 100%;
+    cursor: pointer;
+}
+
+.melofi__scene-modal-thumbnail img {
+    width: 100%;
+    height: 100%;
+}
+.melofi__scene-modal-thumbnail p {
+    color: var(--color-secondary-white);
+    font-size: 32px;
+    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    text-align: center;
+
+}
+
+.melofi__scene-modal-thumbnail-sounds-container {
+    display: flex;
+    column-gap: 10px;
+    margin-top: 10px;
+
+}
+
+.melofi__scene-modal-thumbnail-sounds-container p {
+    font-size: 16px;
+}
+
+@keyframes scale-up {
+    0% {
+      -webkit-transform: scale(0.5);
+              transform: scale(0.5);
+    }
+    100% {
+      -webkit-transform: scale(1);
+              transform: scale(1);
+    }
+  }
